Fix example for `empty`
We were incorrectly showing that you could `insert` a key that might
already exist in an option. That won't type. Instead we use the correct
`set` to change the value of a key in an option.

This might have been causing more confusion than it seems (given it's a
three character change), but it's important to fix these sorts of things.