Add `getAll` to grab every key in an option
As requested in https://github.com/joneshf/purescript-option/issues/2,
we'd like some way to get every value if they all exist. If anything
doesn't exist, we'd like the whole thing to fail.

This is similar in spirit to `Data.Traversable.sequence`, but
monomorphized to `Data.Maybe.Maybe` as the `Applicative`. Since it also
shares quite a bit in common with `get`, we name it accordingly.

This naming hints that there might be more to the family of get* values.
We might think about a way to get a subset of keys from an option. Or,
we might want a way to get a subset of keys with default values.
Something to think about for the future.