Update test suite to use `purescript-spec`
`purescript-assert` is really only useful for one-off low-level
assertions. It doesn't provide any contextual information about what
failed or why.

Since we're writing a package that we expect to maintain for a while, we
want to use a testing package that provides a situation that's a bit
easier to use.