Re-order `Option.ToRecordOption` variables
We want to follow our pattern and make the change easier to address.

There's two parts to this change:
1. Variables that exist only for bookkeeping go toward the start.
2. New variables go at the start.

The first part is that `Option.ToRecordOption _ _` is effectively
`Option.ToRecord`. The idea is that you shouldn't have to bounce around
between variables to really understand what's happening. With the
builder row in the third spot, you have to think of all four variables
of `Option.ToRecordOption` to understand how it relates to
`Option.ToRecord`.

The second part is that whenver you're dealing with a new variable,
there's a boilerplate way to introduce it: at the beginning of the
variables. Every single use can be replaced with a direct
`s/ToRecordOption/ToRecordOption ?value/g`. There's no thinking
required, no clashing to deal with, no parens, etc. When there's some
number of variables to deal with between the `ToRecordOption` and where
the variable gets introduced, it's harder to make the change easier.