Rename `from` to `builder`
We follow our pattern to name the type variable by its use.

When dealing with variables of the same kind, it's hard to know what
they are used for. We've taken to a pattern where we name the type
variable after its use.

I.e. We have the variables `option` and `record` for the `Option.Option
_` row and the `Record` row. We could've just as easily named these `in`
and `out` or some such. We actually named them something similar before
committing, but it was hard to keep track of these things while working
with them.

When we name them after where they're used, it's a bit easier to see why
they exist and what we expect them to do.