Remove record sugar
We continue our pattern of not using the record sugar where possible.

We've been avoiding record sugar as it hides a lot of symmetry. Most of
the actually interesting ideas in this package have come from seeing
some section of code where it was `Record _` rather than `{ | _ }`.

Perhaps it's because we're so adjusted to mentally swapping out English
words. Perhaps it's because the curly braces force us to think
differently. Whatever the reason, we tend to gain more from using less
sugar for records.