Add a way to set from a record
We want to make setting values easier.

There have been a few suggestions and different attempts to make setting
a value easier. What they mostly seem to boil down to is wanting to set
multiple values all at once. While it's possible to do this with a bunch
of different hoops to jump through, we can rethink things to good
effect.

The approach we take here is to take a record of values, and set the
ones that exist. This lets us set a bunch of values at once.