Add a way to set `Data.Maybe.Maybe _` values
We want setting to "Just Work."

We add an instance that can work if you pass in a `Data.Maybe.Maybe _`
value. This makes it a bit more amenable to using in some situation.

This instance is a bit questionable. It works the way we expect when all
goes well. It gives some hard to parse errors when things aren't nested
quite right.

We might have to either improve the instance or remove it alltogether at
some point.