Move `.gitignore` changes to correct section
We don't want to modify the generated section by hand.

The vast majority of the `.gitignore` file is generated. We don't want
to go messing around with the generated section and accidentally lose
some stuff. Instead, we want anything that isn't generated to be in the
`Project-specific` section at the top.

This allows us to re-generate the bottom stuff whenever something
changes.