Move the spirit of the `greet` example to tests
We want larger examples to be in the test suite.

The reason for having a handful of values in the `Option` module is so
we quickly check things that generally cause problems during development
can be caught incredibly quick. The `greet` example was a bit more
complex than we need to check things working correctly.

Instead of removing it entirely, we move the spirit of the example to
the test suite. We'll update the documentation to expand on this example
more when we get around to that point.