Revert behavior of `Option.fromRecord`
We want to make things a bit easier on everyone with these changes.

If we kept `Option.fromRecord` as it was, it would have meant a ton of
breaking changs to anyone that was using that value. Instead of forcing
that on everyone, we revert the behavior to what it was before and
provide a different value that allows required fields.