Don't parameterize builders by a constant
We simplify the type classes since we control all instances.

We always know the first argument is the empty row. We never allow it to
be anything other than that. Nobody else can write an instance because
it'll be an orphan and the language doesn't allow orphan instances.

Instead of letting this builder stuff get out of hand, we simplify
things a bit by explicitly stating what the first argument always is.