Iterate the `Prim.RowList.RowList`
We only need to iterate the `Prim.RowList.RowList` in most instances.

Before, we were iterating both the `Prim.RowList.RowList` and the
`Option.Option _`/`Record _` when implementing the instances. For most
of the instances, we don't actually have to do that.

We can carry around the `Option.Option _`/`Record _` the whole way and
pick pieces off of it as we iterate through the instances. This change
allows us to drop a ton of `Prim.Row.Lacks` constraints, which should be
a net gain overall.