Add a record-vesion of `get`
We want to make it easier to get values.

As with `delete`, `insert`, and `set`, it's a bit cumbersome to pass
around a proxy when we can instead use a record to get better behavior:
we can get multiple values at a time.

We actually make things a bit more flexible with `get'`. Instead of
deciding that the behavior has to be one way or the other, we use
different instances to allow the behavior to work based on what the
value in the supplied record field is: i.e. a function works to map the
value to something else, a `Data.Maybe.Maybe _` works to encode
optionality, and a plain value works as a default. These three types
encode casing on the result, `get`, and `getWithDefault`, respectively.

As with `insert` and `set` dealing with nested `Data.Maybe.Maybe _`s is
not as straight forward as it might seem. The types _can_ guide you
where you want to go, but it takes a minute to get there.