Rename `alter` to `alter'`
We make room for an external export with a non-primed name.

The only reason we've been adding primed names is because we don't want
to unnecessarily make breaking changes. Since `alter` has never been
exposed, we don't have to fall into the same trap. We start off on the
right foot and want to expose the non-primed name with the type we want
to use.