Add `alter`
This is a more expressive version of `modify'`.

Paralleling the other "container" data types (like
`Foreign.Object.Object _`, `Data.Map.Map _ _`, etc.), we want to add a
more expressive version of modifying a value. Here we can say what to do
when a value doesn't exist along with allowing deletion of a value
without also deleting the field.