Build up the resulting `Option _`
We want to insert values instead of assuming a complete `Option _`.

We relaxed the constraints in 0ac03ddb708851b3c5c06a4d2b247eb25776e084.
Unfortunately, we relaxed them a bit too much. If we're returning an
`Option _`, we don't want to assume it's a fully built up `Option _` and
set values on it. We want to start from it having the minimal number of
fields and add them as we build back up.