Copy how-tos to actual PureScript code
We want to automate validating this stuff.

We've had a bunch of how-tos in the README. The thing they lack is a
compiler. To get part of the way there, we copy the code to actual
PureScript modules and test that they work the way we expect them to.

It'd be nice if we didn't have to copy things over like this, but
there's no real facility for making it work otherwise.