Add helper for inserting a field with no value
We don't want to use an untyped escape hatch.

Each time we were casing on the `Option _` and immediately returning a
new `Option _`, it was because we wanted to change the type to have a
new field. We couldn't use `insert` because we had no value to insert.
We ended up constructing a new `Option _` from scratch which allows the
type to be whatever (and have a new field inserted).

The main issue with this is that there's nothing preventing the escape
hatch from escaping too much. I.e. we could change to some type we
really didn't want without knowing it. To mitigate that (and make it so
we're not explicitly destructuring as much), we add a helper that allows
inserting a field with no value.

We could've used `insert'`, but that imposing additional constraints on
each of these type class instances, and we don't need to do that.