Define records with optional and required fields
We want a way to talk about required fields at the data type level.

For a while, we've been trying to think of how to work with required
fields in an `Option.Option _`:
https://github.com/joneshf/purescript-option/issues/12. After having
dealt with it more in anger, it seems like keeping track of which fields
are required might be the way forward:
https://github.com/joneshf/purescript-option/issues/12#issuecomment-688545293.

We added `fromRecordWithRequired` a while back:
https://github.com/joneshf/purescript-option/pull/16. This worked, but
made it even more clear that what we want is a way to carry around
optional and required fields in a data type (rather than only when
calling a function).

We tried changing `Option.Option _` to have required fields on itself.
I.e. `Option.Option _ _`. But that made everything break and was taking
far too long to prove out whether the idea had legs.

We take a different approach of defining another data type that has both
optional and required fields. We name it `Option.Record _ _` because
it's like a record in many other languages, just not a PureScript-style
record. This approach allows us to move forward incrementally instead of
scorching the earth.

We'll have to add more behavior to make working with `Option.Record _ _`
not painful. For now, we add a couple helpers to pull out the underlying
`Option.Option _` and `Record _`. It means that everything is much more
verbose right now, but it at least allows problems to be solved.

We also deprecate `fromRecordWithRequired`. Working with an
`Option.Record _ _` should subsume all that `fromRecordWithRequired` did
before and be easier to use. We'll remove it in 5.0.0 so as not to break
everything when we make this next release.