Derive `Eq _` and `Ord _` instances
We can derive these since they don't give away the structure.

Unlike most of the other instances, `Eq _` and `Ord _` tell nothing
about the underlying structure. This is different from something like
`Show _` or `Data.Argonaut.Decode.Class.DecodeJson _`. Those instances
will tell all of our internals if we were to derive an instance for it.

Since `Eq _` and `Ord _` tell nothing, we derive their instances and
rely on the behavior of the underlying data types to do the right thing.