Hold onto more structure for `encodeJsonOption`
We don't need to go all the way to `Data.Argonaut.Core.Json`.

Instead of losing all structure and returning an opaque
`Data.Argonaut.Core.Json`, we can hold onto the structure of a
`Foreign.Object.Object _` in `encodeJsonOption`. Once we've built the
whole thing, we can wrap it into a `Data.Argonaut.Core.Json` in the
`encodeJsonOptionOption` instance.

The difference is minimal now, but makes quite a difference when we want
to build upon the results of `encodeJsonOption` for `Record _ _`. It
also helps us follow the pattern a bit better that these (should be)
implementation type classes return more structured data that is useful
to work with after the fact.