Add `DecodeJson _` and `EncodeJson _` instances
We want to be able to use `purescript-argonaut` with `Record _ _`.

We add a couple of instances, write a how-to, and codify it as a test.
This is effectively what we did for working with only optional fields.

The instances don't really do a whole lot of unexpected things. Probably
the only interesting bit is that we don't force the underlying structure
of `Option.Record _ _` to be the same as what we decode/encode. If
effect, we pretend that the structure of `Option.Record _ _` is
flattened.