Add `ReadForeign _` and `WriteForeign _` instances
We want to be able to use `purescript-simple-json` with `Record _ _`.

We add a couple of instances, write a how-to, and codify it as a test.
This is effectively what we did for working with only optional fields.

As with the `purescript-argonaut` and `purescript-codec-argonaut` stuff,
we pretend that there's not an extra layer of nesting.