Add `recordSet`
We want to be able to set values in an `Option.Record _ _`.

One of the biggest pain points with using `Option.Option _` when it
first started was having an easy way to set values. Once we added
`set'`, things became considerably easier to work with.

We follow suit for `Option.Record _ _`. We add a way to set values so
it's a bit easier to work with an `Option.Record _ _` after the fact.