Allow `Data.Maybe.Maybe _` values for `fromRecord`
We want to provide an easier API like we did for `set'`.

One of the nice things about `set'` is that we can pass in a value
wrapped in a `Data.Maybe.Maybe _` and it will do the right thing. It
makes the day to day usage much easier since we don't have to worry
about the casing outside of calling `Option.set'`.

We follow suit here and make it easier to construct an `Option.Option _`
from scratch by allowing the record to contain values that might be
wrapped in a `Data.Maybe.Maybe _`.