Bump `purescript-argonaut-codecs` to 7.0.0
This version had a bunch of breaking changes we had to deal with.

The primary thing that changed is the type of decoding. Since we
implement some stuff directly depending on those types, we also
introduce breaking changes.

We also deal with the tests in a way that should mitigate how much we
have to deal with in the future.