Allow setting all fields
We add missing instances to `Option.Partition _ _ _ _ _`.

When we added `Option.Partition _ _ _ _ _`, we didn't make sure that we
could set any field no matter what it was named. We only validated that
it worked on an `Option.Option _`/`Option.Record _ _` with one or two
fields.

Because we didn't have large enough examples, we weren't actually
iterating the `Prim.RowList.RowList _ _ _` enough. In particular, if the
set of fields being set weren't all alphabetically the first fields, we
wouldn't have an instance for `Option.Partition _ _ _ _ _`.

We add two more instances for `Option.Partition _ _ _ _ _`. One that
continues iterating the required fields if they don't match, and one
that continues iterating the optional fields if they don't match. These
two instances allow us to access the rest of the fields in an
`Option.Option _`/`Option.Record _ _`.

Because rows are sorted whenever working with them, this ought to work
out okay. I.e. there _shouldn't_ be an interleaving of required/optional
fields that would be missed by these instances now. If there is, we'll
have to deal with things a little differently.