Fix behavior of `Option.set'`/`Option.recordSet`
We want to remove values when they're `Data.Maybe.Nothing`.

The previous behavior is unintuitive and has caught people off guard.
Instead of keeping it, we make it more consistent with how other
record-based operations work: If the value is `Data.Maybe.Nothing`, we
remove the previous value (if it was set).

Hopefully, this is more intuitive to everyone.