Support null values when parsing optional fields
We want to be more lenient in what we accept.

Up until now, we've required that the JSON we parse come in a specific
format: optional values don't exist. Unfortunately, things aren't so
consistent in the real world.

Instead of failing if we see an optional field that has a `null` value,
we parse it as though the value coming in was `Data.Maybe.Nothing`. This
is akin to how record-based values work (like `Option.set'`).

Since this changes the semantics of parsing, this is a breaking change,
and we should deal with it as such.

With this change, we ought to be able to be used in more places without
causing undue burden.