Bump `purescript` to 0.14.x
We move to the latest version of everything.

We're addressing both errors and warnings in this change (because we'd
like to treat warnings like errors anyway). For the most part this is
straight-forward find/replace change. We did have to do something about
the static checks though.

What we did was dump them all into an array and export that single
value. Since there doesn't seem to be any way to define a value at the
top-level that is otherwise unused without it also producing a warning,
we're kind of pigeon-holed into this sort of solution. The only other
option is to remove the static checks, but they've caught way more
issues than is reasonable for the amount of code they are so that feels
like a major step backwards.

It would've been ideal to prefix them with an `_` (like you can do in
other situations), but that's not how things work. For more information,
see this discussion:
https://github.com/purescript/purescript/pull/3819#issuecomment-725344010.