Bump purescript to 0.15.x
We move to the latest version of everything.

Since v0.15.15 is the most recent stable version at the time of writing,
that's the one we use.

Making this update required some minor implementation changes to get the
types to line up, in particular the change from `Data.Codec.GCodec` to
`Data.Codec.Codec`.