Setup `nix` and `direnv`
We want to get back into developing in this repository. The issue is
that things are so out of date that it's hard to do anything. We do what
we should've done years ago, and setup `nix` and `direnv` here so that
we can at least build stuff.

This is mostly temporary until we get up to date on versions of things.
So even if this doesn't work too well as-is, it hopefully won't be
around for long.