Support ARM platforms in `nix`
Times have moved on, and devices are using ARM more and more these days.
In order to support that in this repository with the version of things
we're using, we have to do a bit of work.

The version of `purescript` we were using–0.14.0–does not have a
pre-built ARM version. The `npm` package for `purescript` will fallback
to building from source, but we have to make sure certain dependencies
are available in order to do that. Since those dependencies are only
necessary if we have to build from source, we make them conditional on
the host platform of the system.

Unfortunately even with those dependencies, we still cannot build things
from source. The version of GHC that `purescript` 0.14.0 wants is
8.10.4, which is not available in the version of `stack`? we're using.
We could go down to an older version of `nixpkgs`, but the alternative
we decide on is to update the version of `purescript` to 0.14.9. This
version of `purescript` requires a newer version of GHC that actually is
available in the `stack`? we're using.

All this stuff means we can now develop again in the year 2025!