Use a poly-kinded `Proxy`
Now that we're on a version of PureScript that supports poly-kinded
types, we don't even have to think about `Data.Symbol.SProxy _`,
`Type.Data.RowList.RLProxy _`, or whatever the next iteration of that
stuff is. We can instead make our internal `Option.Proxy` poly-kinded,
and not deal with the shuffling from `purescript-prelude`.

We update the tests to use `Type.Proxy.Proxy _`, since the tests are
supposed to be validating the public API and making sure it works with
the larger PureScript ecosystem.