Check lints and formatting with `nix`
We move over to using `nix` for formatting and linting. We also setup
some pre-commit and pre-push hooks to validate this stuff.

We only add the formatters and linters that pass without requiring a
change. We'll add more stuff in separate changes, so we can have a
smaller, more manageable diff.

Unfortunately (or maybe fortunately?) we cannot use `purty` anymore. The
developer *cough* hasn't made ARM binaries available, and we're not
going to try to figure out how to build it. More than that, it doesn't
support 0.15.x syntax. We're going to move on to the currently
maintained PureScript formatter in a future change. For now, we remove
it from the `Makefile`.

We also add a `lint` recipe in the `Makefile`. This gets added as a
dependency to `test`, since we generally want to make sure that the
lints pass when testing as well.