Fix compiler warnings
There's two things we fix in this change:
- Ignore third-party dependency warnings.
- Remove erroneous `compile` flag.

We cannot do anything about what some third-party dependency is doing in
this repository. There's no need to surface their warnings to us, so we
ignore them.

We also were erroneously passing `compile` to `psa`. That isn't how that
command works. So we remove `compile`.