Migrate to `purs-tidy`
Since `purty` is deprecated it in favor of `purs-tidy` (I did mention
this in like the old PS Slack or something at some point, but should
probably make it explicit in the repository) we migrate over to
`purs-tidy`.

There are some minor differences in formatting, but that doesn't really
matter. Formatting has never been about doing things "my way." But about
having _something_ format the code so we Humans don't have to think
about it.